Improve Code Separation
