+++ projects/com.oracle.truffle.llvm.parser.base/src/com/oracle/truffle/llvm/parser/base/model/ModelModule.java
@@ -40,16 +40,16 @@ import com.oracle.truffle.llvm.parser.base.model.globals.GlobalValueSymbol;
 import com.oracle.truffle.llvm.parser.base.model.globals.GlobalVariable;
 import com.oracle.truffle.llvm.parser.base.model.symbols.Symbol;
 import com.oracle.truffle.llvm.parser.base.model.symbols.Symbols;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.BigIntegerConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.integer.BigIntegerConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.BinaryOperationConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.BlockAddressConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.CastConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.CompareConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.Constant;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.FloatingPointConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.floatingpoint.FloatingPointConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.GetElementPointerConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.InlineAsmConstant;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.IntegerConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.integer.IntegerConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.NullConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.StringConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.UndefinedConstant;
