+++ projects/com.oracle.truffle.llvm.parser.bc.impl/src/com/oracle/truffle/llvm/parser/bc/impl/nodes/LLVMNodeGenerator.java
@@ -80,21 +80,21 @@ import com.oracle.truffle.llvm.parser.base.model.functions.FunctionParameter;
 import com.oracle.truffle.llvm.parser.base.model.globals.GlobalValueSymbol;
 import com.oracle.truffle.llvm.parser.base.model.symbols.Symbol;
 import com.oracle.truffle.llvm.parser.base.model.symbols.ValueSymbol;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.ArrayConstant;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.BigIntegerConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.aggregate.ArrayConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.integer.BigIntegerConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.BinaryOperationConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.BlockAddressConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.CastConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.CompareConstant;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.FloatingPointConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.floatingpoint.FloatingPointConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.GetElementPointerConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.InlineAsmConstant;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.IntegerConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.integer.IntegerConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.MetadataConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.NullConstant;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.StructureConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.aggregate.StructureConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.constants.UndefinedConstant;
-import com.oracle.truffle.llvm.parser.base.model.symbols.constants.VectorConstant;
+import com.oracle.truffle.llvm.parser.base.model.symbols.constants.aggregate.VectorConstant;
 import com.oracle.truffle.llvm.parser.base.model.symbols.instructions.ValueInstruction;
 import com.oracle.truffle.llvm.parser.base.model.enums.AsmDialect;
 import com.oracle.truffle.llvm.parser.base.model.enums.BinaryOperator;
