+++ projects/com.oracle.truffle.llvm.parser.bc.impl/src/com/oracle/truffle/llvm/parser/bc/impl/parser/bc/Parser.java
@@ -38,6 +38,7 @@ import com.oracle.truffle.llvm.parser.bc.impl.parser.bc.blocks.Block;
 import com.oracle.truffle.llvm.parser.bc.impl.parser.bc.records.UserRecordArrayOperand;
 import com.oracle.truffle.llvm.parser.bc.impl.parser.bc.records.UserRecordBuilder;
 import com.oracle.truffle.llvm.parser.bc.impl.parser.bc.records.UserRecordOperand;
+import com.oracle.truffle.llvm.parser.bc.impl.parser.listeners.ParserListener;
 import com.oracle.truffle.llvm.parser.bc.impl.parser.util.Pair;
 
 public class Parser {
