+++ projects/com.oracle.truffle.llvm.parser.bc.impl/src/com/oracle/truffle/llvm/parser/bc/impl/parser/listeners/metadata/Metadata.java
@@ -27,7 +27,7 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.oracle.truffle.llvm.parser.bc.impl.parser.ir.module;
+package com.oracle.truffle.llvm.parser.bc.impl.parser.listeners.metadata;
 
 import java.util.Arrays;
 import java.util.List;
@@ -54,10 +54,11 @@ import com.oracle.truffle.llvm.parser.base.model.metadata.MetadataSubprogram;
 import com.oracle.truffle.llvm.parser.base.model.metadata.MetadataSubrange;
 import com.oracle.truffle.llvm.parser.base.model.metadata.MetadataValue;
 import com.oracle.truffle.llvm.parser.base.model.types.Type;
+import com.oracle.truffle.llvm.parser.bc.impl.parser.listeners.ParserListener;
+import com.oracle.truffle.llvm.parser.bc.impl.parser.listeners.Types;
 import com.oracle.truffle.llvm.runtime.LLVMLogger;
 import com.oracle.truffle.llvm.runtime.options.LLVMBaseOptionFacade;
 
-import com.oracle.truffle.llvm.parser.bc.impl.parser.bc.ParserListener;
 import com.oracle.truffle.llvm.parser.bc.impl.parser.ir.module.records.MetadataRecord;
 
 public class Metadata implements ParserListener {
