+++ projects/com.oracle.truffle.llvm.parser.bc.impl/src/com/oracle/truffle/llvm/parser/bc/impl/parser/listeners/module/ModuleV38.java
@@ -27,9 +27,10 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.oracle.truffle.llvm.parser.bc.impl.parser.ir.module;
+package com.oracle.truffle.llvm.parser.bc.impl.parser.listeners.module;
 
 import com.oracle.truffle.llvm.parser.base.model.generators.ModuleGenerator;
+import com.oracle.truffle.llvm.parser.bc.impl.parser.listeners.ModuleVersion;
 
 public class ModuleV38 extends Module {
